requirements.txt
# Web Server & Framework
flask~=3.1.2
gunicorn

# Data & Finance
numpy~=2.1.3
pandas~=2.2.0
yfinance~=0.2.50

# API & Environment
requests~=2.32.5
python-dotenv~=1.1.0
google-genai

# Sentiment & Logic
vadersentiment~=3.3.2
protobuf~=5.29.3